+++ app/src/renderer/vuex/modules/governance/proposals.js
@@ -12,6 +12,9 @@ export default ({ node }) => {
   const mutations = {
     setProposal(state, proposal) {
       Vue.set(state.proposals, proposal.proposal_id, proposal)
+    },
+    setProposalTally(state, proposalId, tally) {
+      state.proposals[proposalId].tally_result = tally
     }
   }
   let actions = {
@@ -32,6 +35,13 @@ export default ({ node }) => {
         if (proposals.length > 0) {
           proposals.forEach(proposal => {
             commit(`setProposal`, proposal.value)
+            // the proposal doesn't hold the tally results until it's inactive (rejected or passed)
+            // TODO: enable after upgrading to latest SDK
+            // if (proposal.value.proposal_status === `VotingPeriod`) {
+            //   node.queryProposalTally(proposal.value.proposal_id).then(tally => {
+            //     commit(`setProposalTally`, proposal.value.proposal_id, tally)
+            //   })
+            // }
           })
         }
       } catch (error) {
@@ -55,6 +65,7 @@ export default ({ node }) => {
           title: `Error querying proposal with id #${proposal_id}`,
           body: error.message
         })
+        Raven.captureException(error)
         state.error = error
       }
       state.loading = false